Refactor: custom_app_bar.dart로 파일 이름 변경
